fixed create

+++ src/main/resources/sql/dbCreate-v2.sql
@@ -37,7 +37,7 @@ CREATE TABLE WorkPackage (
   estimateCost numeric(20, 4), 
   actualCost   numeric(20, 4), 
   parentwpID   int(10),
-  status       int(2),
+  status       tinyint(1),
   CONSTRAINT packageID
     PRIMARY KEY (packageID));
 --
@@ -48,8 +48,8 @@ CREATE TABLE Project (
   startDate    date NOT NULL, 
   endDate      date, 
   budget       numeric(20, 4), 
-  status       int(2), 
-  pmEmployeeID varchar(10),
+  status       tinyint(1), 
+  pmEmployeeID    varchar(10),
   genReport       tinyint(1) NOT NULL,
   CONSTRAINT projectID  
     PRIMARY KEY (projectID)
